feat: 초기 프로젝트 설정
