+++ .env.example
@@ -1,0 +1,10 @@
+# Database
+DATABASE_URL=sqlite:///./blog.db
+
+# Security
+SECRET_KEY=replace-with-your-secret
+
+
+# JWT settings (선택)
+ALGORITHM=HS256
+ACCESS_TOKEN_EXPIRE_MINUTES=30
