+++ back/app/models/user.py
@@ -1,0 +1,9 @@
+from sqlalchemy import Column, Integer, String
+from back.app.core.database import Base
+
+class User(Base):
+    __tablename__ = "users"
+
+    id = Column(Integer, primary_key=True, index=True)
+    username = Column(String, unique=True, index=True, nullable=False)
+    password = Column(String, nullable=False)
